LayoutTests/css3/filters/custom-with-at-rule-syntax/parsing-parameters-property-valid-expected.txt

Test parameters property parsing in the @-webkit-filter at-rule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



========================================
Number parameter tests.
========================================

Integer parameters.
parameters: n1 1, n2 2 3, n3 3 4 5, n4 4 5 6 7;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "n1 1, n2 2 3, n3 3 4 5, n4 4 5 6 7"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 4
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "n1 1"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "n2 2 3"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "n3 3 4 5"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "n4 4 5 6 7"

Float parameters.
parameters: n1 1.1, n2 2.2 3.3, n3 3.1 4.1 5.1, n4 4.1 5.2 6.3 7.4;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "n1 1.1, n2 2.2 3.3, n3 3.1 4.1 5.1, n4 4.1 5.2 6.3 7.4"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 4
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "n1 1.1"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "n2 2.2 3.3"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "n3 3.1 4.1 5.1"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "n4 4.1 5.2 6.3 7.4"

Parameter name same as CSS keyword.
parameters: background 0 1 0 1;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "background 0 1 0 1"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "background 0 1 0 1"

========================================
Transform parameter tests.
========================================

Transform parameter with one transform function.
parameters: t rotate(0deg);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "t rotate(0deg)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "t rotate(0deg)"

Transform parameter with multiple transform functions.
parameters: t rotate(0deg) perspective(0) scale(0, 0) translate(0px, 0px);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "t rotate(0deg) perspective(0) scale(0, 0) translate(0px, 0px)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "t rotate(0deg) perspective(0) scale(0, 0) translate(0px, 0px)"

Mulitple transform parameters.
parameters: t1 rotate(0deg), t2 rotate(0deg);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "t1 rotate(0deg), t2 rotate(0deg)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 2
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "t1 rotate(0deg)"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "t2 rotate(0deg)"

========================================
Array parameter tests.
========================================

Array parameter with name 'array'.
parameters: array array(1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "array array(1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "array array(1)"

Array parameter with one positive integer.
parameters: a array(1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "a array(1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "a array(1)"

Array parameter with one negative float.
parameters: a array(-1.01);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "a array(-1.01)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "a array(-1.01)"

Array parameter with multiple positive integers.
parameters: a array(1, 2, 3, 4, 5);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "a array(1, 2, 3, 4, 5)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "a array(1, 2, 3, 4, 5)"

Array parameter with multiple signed floats.
parameters: a array(1, -2.2, 3.14, 0.4, 5);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "a array(1, -2.2, 3.14, 0.4, 5)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "a array(1, -2.2, 3.14, 0.4, 5)"

Multiple array parameters.
parameters: a1 array(1, -2.2, 3.14, 0.4, 5), a2 array(1, 2, 3);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "a1 array(1, -2.2, 3.14, 0.4, 5), a2 array(1, 2, 3)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 2
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "a1 array(1, -2.2, 3.14, 0.4, 5)"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "a2 array(1, 2, 3)"

========================================
Color parameter tests.
========================================

Hex color.
parameters: c #00FF00;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "c rgb(0, 255, 0)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "c rgb(0, 255, 0)"

Color keyword.
parameters: c green;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "c rgb(0, 128, 0)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "c rgb(0, 128, 0)"

Color rgb function.
parameters: c rgb(0, 128, 0);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "c rgb(0, 128, 0)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "c rgb(0, 128, 0)"

Color hsl function.
parameters: c hsl(120, 100%, 50%);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "c rgb(0, 255, 0)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "c rgb(0, 255, 0)"

Color rgba function.
parameters: c rgba(0, 255, 0, 0.2);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "c rgba(0, 255, 0, 0.2)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "c rgba(0, 255, 0, 0.2)"

Color hsla function.
parameters: c hsla(120, 100%, 50%, 0.2);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "c rgba(0, 255, 0, 0.2)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "c rgba(0, 255, 0, 0.2)"

========================================
Mat2 parameter tests.
========================================

Mat2 parameter.
parameters: m mat2(1, 0, 0, 1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat2(1, 0, 0, 1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat2(1, 0, 0, 1)"

Mat2 parameter with negative values.
parameters: m mat2(-1, -1, -1, -1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat2(-1, -1, -1, -1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat2(-1, -1, -1, -1)"

Mat2 parameter with negative and positive values.
parameters: m mat2(-1, 1, 1, -1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat2(-1, 1, 1, -1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat2(-1, 1, 1, -1)"

Mat2 parameter with multiple signed floats.
parameters: m mat2(1, -2.2, 3.14, 0.4);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat2(1, -2.2, 3.14, 0.4)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat2(1, -2.2, 3.14, 0.4)"

========================================
Mat3 parameter tests.
========================================

Mat3 parameter.
parameters: m mat3(1, 0, 0, 0, 1, 0, 0, 0, 1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat3(1, 0, 0, 0, 1, 0, 0, 0, 1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat3(1, 0, 0, 0, 1, 0, 0, 0, 1)"

Mat3 parameter with negative values.
parameters: m mat3(-1, -1, -1, -1, -1, -1, -1, -1, -1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat3(-1, -1, -1, -1, -1, -1, -1, -1, -1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat3(-1, -1, -1, -1, -1, -1, -1, -1, -1)"

Mat3 parameter with negative and positive values.
parameters: m mat3(-1, 1, 1, -1, -1, 1, 1, -1, 1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat3(-1, 1, 1, -1, -1, 1, 1, -1, 1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat3(-1, 1, 1, -1, -1, 1, 1, -1, 1)"

Mat3 parameter with multiple signed floats.
parameters: m mat3(1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4, 1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat3(1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4, 1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat3(1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4, 1)"

========================================
Mat4 parameter tests.
========================================

Mat4 parameter.
parameters: m mat4(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat4(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat4(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1)"

Mat4 parameter with negative values.
parameters: m mat4(-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat4(-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat4(-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1)"

Mat4 parameter with negative and positive values.
parameters: m mat4(-1, 1, 1, -1, -1, 1, 1, -1, -1, 1, 1, -1, -1, 1, 1, -1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat4(-1, 1, 1, -1, -1, 1, 1, -1, -1, 1, 1, -1, -1, 1, 1, -1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat4(-1, 1, 1, -1, -1, 1, 1, -1, -1, 1, 1, -1, -1, 1, 1, -1)"

Mat4 parameter with multiple signed floats.
parameters: m mat4(1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "m mat4(1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 1
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m mat4(1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4, 1, -2.2, 3.14, 0.4)"

========================================
Combined parameter tests.
========================================

Number parameter, color parameter, transform parameter, matrix parameters and array parameter.
parameters: n 1, c rgb(0, 128, 0), t rotate(0deg), m1 mat2(1, 0, 0, 1), m2 mat3(1, 0, 0, 0, 1, 0, 0, 0, 1), m3 mat4(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1), a array(1);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 1
PASS removeBaseURL(styleDeclaration.getPropertyValue('parameters')) is "n 1, c rgb(0, 128, 0), t rotate(0deg), m1 mat2(1, 0, 0, 1), m2 mat3(1, 0, 0, 0, 1, 0, 0, 0, 1), m3 mat4(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1), a array(1)"
PASS parametersPropertyValue instanceof CSSValueList is true
PASS parametersPropertyValue.constructor === CSSValueList is true
PASS parametersPropertyValue.__proto__ === CSSValueList.prototype is true
PASS parametersPropertyValue.length is 7
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "n 1"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "c rgb(0, 128, 0)"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "t rotate(0deg)"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m1 mat2(1, 0, 0, 1)"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m2 mat3(1, 0, 0, 0, 1, 0, 0, 0, 1)"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "m3 mat4(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1)"
PASS subValue.cssValueType is CSSValue.CSS_VALUE_LIST
PASS removeBaseURL(subValue.cssText) is "a array(1)"
PASS successfullyParsed is true

TEST COMPLETE
